Extract profile page to user module
